+++ src/containers/Main.js
@@ -1,10 +1,11 @@
 import { connect } from 'react-redux';
 import { Main } from '../components';
-import { usersToggler } from '../controller/actions';
+import { usersToggler, hideSidebar } from '../controller/actions';
 
 const mapDispatchToState = state => {
   return ({
     users: state.users.list,
+    mobileLayout: state.users.mobileLayout,
     showUsersList: state.users.mobileLayout,
     connectionStatus: state.currentUser.connectionStatus,
     currentChat: state.currentChat,
@@ -13,7 +14,8 @@ const mapDispatchToState = state => {
 };
 
 const mapDispatchStateToProps = dispatch => ({
-  usersListToogle: () => dispatch(usersToggler())
+  usersListToogle: () => dispatch(usersToggler()),
+  hideSidebar: () => dispatch(hideSidebar())
 });
 
 export default connect(mapDispatchToState, mapDispatchStateToProps)(Main);