+++ src/controller/actions/index.js
@@ -1,5 +1,5 @@
 import { initialLoad } from './common';
-import { addUser, populateUsersList, userInitializationStart, userInitializationFinish, userDisconected, usersToggler } from './users';
+import { addUser, populateUsersList, userInitializationStart, userInitializationFinish, userDisconected, usersToggler, hideSidebar } from './users';
 import { addMessage, messageReceived, liveTyping, userTyping } from './messages';
 import { errorReceived, showDebug, hideDebug, hideDebugManual } from './debug';
 import { selectChat, chatSelected, createNewDirectChat, chatCreated } from './chats';
@@ -18,6 +18,7 @@ export {
   hideDebug,
   hideDebugManual,
   usersToggler,
+  hideSidebar,
   liveTyping,
   userTyping,
   selectChat,